Switched spawn system back to using ints, added a max attempts, as
well as fixed using the wrong node list in the despawn method

+++ src/microbe_stage/ISpawned.cs
@@ -9,7 +9,7 @@ public interface ISpawned
     ///   If the squared distance to the player of this object is
     ///   greater than this, it is despawned.
     /// </summary>
-    float DespawnRadiusSqr { get; set; }
+    int DespawnRadiusSqr { get; set; }
 
     /// <summary>
     ///   The Node that was spawned for this entity, used to tag it
